requirements.txt
APScheduler==3.10.1
async-timeout==4.0.2
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
docopt==0.4.0
Flask==2.2.3
Flask-APScheduler==1.12.4
Flask-SocketIO==5.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mailchimp-transactional==1.0.50
mandrill==1.0.60
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
requests==2.28.2
secure-smtplib==0.1.1
six==1.16.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.2.3
